+++ src/Raytracing/Primitives/Primitive.h
@@ -29,8 +29,8 @@ namespace Raycer
 		virtual AABB getAABB() const = 0;
 		virtual void transform(const Vector3& scale, const EulerAngle& rotate, const Vector3& translate) = 0;
 
-		size_t id = 0;
-		size_t materialId = 0;
+		uint64_t id = 0;
+		uint64_t materialId = 0;
 		bool invisible = false;
 		Material* material = nullptr;
