+++ src/Rendering/Image.h
@@ -20,18 +20,18 @@ namespace Raycer
 	public:
 
 		Image();
-		Image(size_t length);
-		Image(size_t width, size_t height);
-		Image(size_t width, size_t height, float* rgbaData);
+		Image(uint64_t length);
+		Image(uint64_t width, uint64_t height);
+		Image(uint64_t width, uint64_t height, float* rgbaData);
 		Image(const std::string& fileName);
 
-		void load(size_t width, size_t height, float* rgbaData);
+		void load(uint64_t width, uint64_t height, float* rgbaData);
 		void load(const std::string& fileName);
 		void save(const std::string& fileName) const;
-		void resize(size_t length);
-		void resize(size_t width, size_t height);
-		void setPixel(size_t x, size_t y, const Color& color);
-		void setPixel(size_t index, const Color& color);
+		void resize(uint64_t length);
+		void resize(uint64_t width, uint64_t height);
+		void setPixel(uint64_t x, uint64_t y, const Color& color);
+		void setPixel(uint64_t index, const Color& color);
 		void clear();
 		void clear(const Color& color);
 		void applyFastGamma(double gamma);
@@ -39,12 +39,12 @@ namespace Raycer
 		void flip();
 		void fillTestPattern();
 
-		size_t getWidth() const;
-		size_t getHeight() const;
-		size_t getLength() const;
+		uint64_t getWidth() const;
+		uint64_t getHeight() const;
+		uint64_t getLength() const;
 
-		Color getPixel(size_t x, size_t y) const;
-		Color getPixel(size_t index) const;
+		Color getPixel(uint64_t x, uint64_t y) const;
+		Color getPixel(uint64_t index) const;
 		Color getPixelNearest(double u, double v) const;
 		Color getPixelBilinear(double u, double v) const;
 
@@ -53,8 +53,8 @@ namespace Raycer
 
 	private:
 
-		size_t width = 0;
-		size_t height = 0;
+		uint64_t width = 0;
+		uint64_t height = 0;
 
 		AlignedColorfVector pixelData;
 	};
